+++ src/rinq/internal/optutil/ginkgo_test.go
@@ -1,0 +1,13 @@
+package optutil_test
+
+import (
+	"testing"
+
+	"github.com/onsi/ginkgo"
+	"github.com/onsi/gomega"
+)
+
+func TestSuite(t *testing.T) {
+	gomega.RegisterFailHandler(ginkgo.Fail)
+	ginkgo.RunSpecs(t, "optutil")
+}