+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Next Release
+
+- **[BC]** Remove `Config` in favour of "functional options" in the `options` package
+
 ## 0.4.0 (2017-09-18)
 
 Please note that this release includes changes to the definition of AMQP